Revised draft at minimizing use of stream adapters
